client/pyproject.toml
[tool.poetry]
name = "termchat"
version = "1.0.0"
description = ""
authors = ["jayden <jayden@jaydensar.net>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
requests = "^2.25.1"
blessed = "^1.18.0"
python-dotenv = "^0.17.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
